Revert "MAR-2670. New sitemap structure"
