+++ server/config/json/mainRedirects.json
@@ -1020,11 +1020,6 @@
   },
   {
     "pageType": "Tags",
-    "from": "/blog/project-management",
-    "to": "https://maddevs.io/tag/project-management/"
-  },
-  {
-    "pageType": "Tags",
     "from": "/insights/blog/tag/mobile-development",
     "to": "https://maddevs.io/tag/mobile-app-development/"
   },
